Added .sh files for 16bit length and 64bit length. Added the output text files. Cheterodyning_threaded.c is configured for 64 bit, CHeterodyning.c is still configured is 16 bit and globals.h now has capabilities for 16 bit but the change doesn't affect 32 or 64 bit

+++ C/src/CHeterodyning_threaded.c
@@ -1,6 +1,6 @@
 #include "CHeterodyning_threaded.h"
 
-float result [SAMPLE_COUNT];
+double result [SAMPLE_COUNT];
 
 // This is each thread's "main" function.  It receives a unique ID
 void* Thread_Main(void* Parameter){
